The modifications of this commit include: change of directory organization and the creation of a script to handle the farm_branches.csv database.
The project now has a "Source" directory and a "Data Processing" directory. A new script called "treatment_farm_branches.py" has been created to handle the data from the "Farm_Branches" dataset. The script focuses on converting geographical coordinates from degree, minute, and second format to decimal values. The steps of this treatment are outlined in the "data_processing_notes.txt" file. The processed data is saved in the "cg_farm_branches_to_decimal.csv" file as the final result.